Refactor : 테스트 코드 & 날씨 단기예보 임시 추가 & 경위도 변환기 추가
