+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/ui/detail/DetailViewModel.kt
@@ -21,8 +21,7 @@ import com.wannabeinseoul.seoulpublicservice.pref.IdPrefRepository
 import com.wannabeinseoul.seoulpublicservice.pref.RecentPrefRepository
 import com.wannabeinseoul.seoulpublicservice.pref.SavedPrefRepository
 import com.wannabeinseoul.seoulpublicservice.ui.dialog.review.ReviewItem
-import com.wannabeinseoul.seoulpublicservice.weather.shorttime.WeatherShortDTO
-import com.wannabeinseoul.seoulpublicservice.weather.shorttime.WeatherShortRepository
+import com.wannabeinseoul.seoulpublicservice.weather.WeatherShortRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
@@ -69,8 +68,8 @@ class DetailViewModel(
     private val _favoriteChanged: MutableLiveData<Boolean> = MutableLiveData()
     val favoriteChanged: LiveData<Boolean> get() = _favoriteChanged
 
-    private val _shortWeather: MutableLiveData<WeatherShortDTO> = MutableLiveData()
-    val shortWeather: LiveData<WeatherShortDTO> get() = _shortWeather
+//    private val _shortWeather: MutableLiveData<WeatherShortDTO> = MutableLiveData()
+//    val shortWeather: LiveData<WeatherShortDTO> get() = _shortWeather
 
     fun getData(svcID: String) {
         viewModelScope.launch{
@@ -186,11 +185,11 @@ class DetailViewModel(
         recentPrefRepository.setRecent(data)
     }
 
-    fun getWeather(page: Int, row: Int, date: String, time: String, x: Int, y: Int) {
-        runBlocking(Dispatchers.IO) {
-            weatherShortRepository.getShortWeather(page, row, date, time, x, y)
-        }.let {
-            _shortWeather.value = it
-        }
-    }
+//    fun getWeather(page: Int, row: Int, date: String, time: String, x: Int, y: Int) {   // 테스트용
+//        runBlocking(Dispatchers.IO) {
+//            weatherShortRepository.getShortWeather(page, row, date, time, x, y)
+//        }.let {
+//            _shortWeather.value = it
+//        }
+//    }
 }
