+++ app/src/main/res/layout/fragment_home.xml
@@ -131,6 +131,16 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/view_pager" />
 
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/rv_home_weather_week"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="8dp"
+                app:layout_constraintBottom_toTopOf="@+id/tv_home_recent_title"
+                app:layout_constraintEnd_toEndOf="@+id/iv_home_main_banner"
+                app:layout_constraintStart_toStartOf="@+id/iv_home_main_banner"
+                app:layout_constraintTop_toBottomOf="@+id/iv_home_main_banner" />
+
             <TextView
                 android:id="@+id/tv_home_recent_title"
                 android:layout_width="wrap_content"
@@ -141,7 +151,7 @@
                 android:textSize="16sp"
                 android:textStyle="bold"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/iv_home_main_banner" />
+                app:layout_constraintTop_toBottomOf="@id/rv_home_weather_week" />
 
             <TextView
                 android:id="@+id/tv_home_recent_description"
